fix(ci): resolve mypy errors in core + api; fix Vercel monorepo build
- packages/spectraquant-core: expand mypy overrides to cover
  pandas_market_calendars (top-level) and statsmodels (top-level),
  disable import-untyped error code for both
- apps/api: add jose, structlog, asyncpg, redis, statsmodels,
  pandas_market_calendars to ignore_missing_imports overrides;
  disable import-untyped; jobs module retains ignore_errors
- apps/api/src/jobs/validate_backfill.py: cast sharpe to float()
  to fix 'float <= object' operator error (mypy couldn't narrow
  np.floating through chained comparison)
- vercel.json: switch buildCommand from 'cd apps/web && pnpm build'
  to 'pnpm --filter web exec next build' to prevent Turbo from
  intercepting the Vercel build and trying to build all packages;
  add NEXT_PUBLIC_BACKFILL_START_DATE env var needed at build time

+++ apps/api/src/jobs/validate_backfill.py
@@ -114,7 +114,7 @@ def main(argv: list[str] | None = None) -> int:
             )
             mom = np.array([r["long_short_return"] for r in mom_rows], dtype=float)
             if len(mom) > 20:
-                sharpe = (mom.mean() / mom.std(ddof=1)) * math.sqrt(252)
+                sharpe = float((mom.mean() / mom.std(ddof=1)) * math.sqrt(252))
                 if 0.2 <= sharpe <= 1.5:
                     print(f"PASS [sharpe] momentum Sharpe={sharpe:.2f} ✓")
                 else:
